PencilArrays are accessed with unpermuted indices
Permutations are now performed on the fly. It's now much easier to work
with PencilArrays regardless of their underlying index permutations.

This required a custom implementation of LinearIndices() and of
Base._sub2ind(). I'm not sure if there's a more proper way to get
consistency between linear and Cartesian indices...